+++ .gitignore
@@ -156,3 +156,7 @@ Thumbs.db
 # Common editor files
 *~
 *.swp
+
+# notebooks
+Untitiled.ipynb
+t.ipynb
