+++ examples/nets.py
@@ -8,7 +8,7 @@ import jax.numpy as jnp
 import jax.random as jrandom
 
 from signax.module import SignatureTransform
-from signax.signature import signature, signature_combine
+from signax import signature, signature_combine
 from signax.utils import flatten
