+++ src/signax/__init__.py
@@ -2,4 +2,21 @@ from __future__ import annotations
 
 __version__ = "0.1.2"
 
-__all__ = ("__version__",)
+__all__ = (
+    "__version__",
+    "module",
+    "signature",
+    "logsignature",
+    "signature_combine",
+    "signature_to_logsignature",
+    "multi_signature_combine",
+)
+
+from signax import module
+from signax.signatures import (
+    logsignature,
+    multi_signature_combine,
+    signature,
+    signature_combine,
+    signature_to_logsignature,
+)
