+++ src/signax/module.py
@@ -3,7 +3,8 @@ from __future__ import annotations
 import equinox as eqx
 import jax
 
-from signax.signature_flattened import signature, signature_combine
+from signax import signature, signature_combine
+from signax.utils import flatten, unravel_signature
 
 
 class SignatureTransform(eqx.Module):
@@ -16,7 +17,7 @@ class SignatureTransform(eqx.Module):
         self,
         path: jax.Array,
     ) -> jax.Array:
-        return signature(path, self.depth)
+        return flatten(signature(path, self.depth))
 
 
 class SignatureCombine(eqx.Module):
@@ -28,4 +29,6 @@ class SignatureCombine(eqx.Module):
         self.depth = depth
 
     def __call__(self, signature1: jax.Array, signature2: jax.Array):
-        return signature_combine(signature1, signature2, self.dim, self.depth)
+        sig1 = unravel_signature(signature1, self.dim, self.depth)
+        sig2 = unravel_signature(signature2, self.dim, self.depth)
+        return flatten(signature_combine(sig1, sig2))
